only the owners can edit event

+++ src/events/events.controller.ts
@@ -13,6 +13,7 @@ import {
   Logger,
   NotFoundException,
   Query,
+  ForbiddenException,
 } from '@nestjs/common';
 import { Repository, MoreThan, Like, Not } from 'typeorm';
 
@@ -126,28 +127,53 @@ export class EventsController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id: string, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOneBy({ id: +id });
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id: string,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventsService.getEvent(+id);
 
     if (!event) {
       this.logger.debug(`Event with ID ${id} not found`);
       throw new NotFoundException();
     }
 
-    return await this.repository.save({
-      ...input,
-      when: input.when ? new Date(input.when) : event?.when,
-    });
+    if (event.organizerId !== user.id) {
+      this.logger.debug(
+        `User ${user.id} is not the organizer of event with ID ${id}`,
+      );
+      throw new ForbiddenException(
+        null,
+        'You are not the organizer of this event',
+      );
+    }
+
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id') id: string) {
-    const result = await this.eventsService.deleteEvent(+id);
+  @UseGuards(AuthGuardJwt)
+  async remove(@Param('id') id: string, @CurrentUser() user: User) {
+    const event = await this.eventsService.getEvent(+id);
 
-    if (result.affected !== 1) {
+    if (!event) {
       this.logger.debug(`Event with ID ${id} not found`);
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      this.logger.debug(
+        `User ${user.id} is not the organizer of event with ID ${id}`,
+      );
+      throw new ForbiddenException(
+        null,
+        'You are not the organizer of this event',
+      );
+    }
+
+    return await this.eventsService.deleteEvent(+id);
   }
 }
